+++ src/compiler/llvm_codegen_builtins.c
@@ -1009,6 +1009,7 @@ void llvm_emit_builtin_call(GenContext *c, BEValue *result_value, Expr *expr)
 		case BUILTIN_STR_FIND:
 		case BUILTIN_WIDESTRING_16:
 		case BUILTIN_WIDESTRING_32:
+		case BUILTIN_RND:
 			UNREACHABLE
 		case BUILTIN_NONE:
 			UNREACHABLE